readers refactoring
