--- docs/api/utils/data.md
@@ -1,7 +1,0 @@
-::: sopa.utils.data.uniform
-    options:
-      show_root_heading: true
-
-::: sopa.utils.data.blobs
-    options:
-      show_root_heading: true
