+++ sopa/io/reader/wsi.py
