+++ workflow/Snakefile
@@ -93,7 +93,7 @@ rule patch_segmentation_baysor:
     params:
         args_baysor_prior_seg = args.baysor_prior_seg,
     resources:
-        mem_mb=64_000,
+        mem_mb=128_000,
     shell:
         """
         if command -v module &> /dev/null; then
